requirements.txt
asgiref==3.3.1
Django==3.1.7
django-cors-headers==3.7.0
django-environ==0.4.5
django-sslserver==0.22
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1

six~=1.15.0